pcos-pcod-ai-project/requirements.txt
Flask
flask-cors
numpy
pandas
scikit-learn
Pillow
tensorflow
joblib
werkzeug
